Added test_mandelbrot to exercise keyboard and mouse Added keyboard and mouse handlers to animwin32

+++ raster_rgba.cpp
@@ -231,7 +231,7 @@ void raster_rgba_blit(pb_rgba *pb, unsigned int x, unsigned int y, pb_rgba *src)
 	// IMPROVE
 	// use _span
 
-	for (size_t srcrow = 0; srcrow < src->frame.height; srcrow++)
+	for (int srcrow = 0; srcrow < src->frame.height; srcrow++)
 	{
 		// for each row of the source
 		// copy to the destination
